// src/component/career/interfaces/lookups.interface.ts
export interface LocationsTypes {
  id: number;
  title: string;
  city: string;
  state: string;
  country: string;
  zip: string;
}

export interface FunctionsTypes {
  id: number;
  title: string;
}

export interface DepartmentsTypes {
  id: number;
  title: string;
}

export interface DivisionsTypes {
  id: number;
  title: string;
}